refactor: gateway에 security 설정 옮기기

See merge request s11-final/S11P31B206!78

+++ backend/homeggu-user/src/main/java/com/homeggu/domain/saledBoardLike/repository/LikeRepository.java
@@ -9,10 +9,10 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface LikeRepository extends JpaRepository<SalesBoardLike, Integer> {
-    Optional<SalesBoardLike> findByUserAndSalesBoardId(User user, int salesBoardId);
+public interface LikeRepository extends JpaRepository<SalesBoardLike, Long> {
+    Optional<SalesBoardLike> findByUserAndSalesBoardId(User user, Long salesBoardId);
 
     List<SalesBoardLike> findByUser(User user);
 
-    boolean existsByUserAndSalesBoardId(User user, int salesBoardId);
+    boolean existsByUserAndSalesBoardId(User user, Long salesBoardId);
 }
