+++ backend/homeggu-user/src/main/java/com/homeggu/domain/userProfile/repository/UserProfileRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
-    Optional<UserProfile> findByUserProfileId(int userProfileId);
+public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
+    Optional<UserProfile> findByUserProfileId(Long userProfileId);
 
     boolean existsByNickname(String nickname);
 }
